move message types around
extract common types to shared package
fix port entry issues on server
