--- Projects/MakeMeLaugh_Client/Assets/Scripts/JokeTempalte.cs
@@ -1,9 +1,0 @@
-﻿public class JokeTempalte
-{
-  public string Setup { get; set; }
-
-  public string Build { get; set; }
-
-  public string Punchline { get; set; }
-
-}
