+++ Projects/MakeMeLaugh_Client/Assets/Scripts/MainUIViewModel.cs
@@ -141,7 +141,6 @@ public class MainUIViewModel : MonoBehaviour
       ConnectionManager.JokesReceived += OnJokeReceived;
     }
 
-    //should we validate befoore closing?
     ClosePopUp(_settingsView);
 
     //go to waiting popup
