+++ SharedPackages/Assets/Scripts/JokeState.cs
@@ -1,0 +1,7 @@
+﻿public enum JokeState
+{
+  Setup,
+  Punchline,
+  Done,
+}
+
