+++ SharedPackages/Assets/Scripts/PlayerPunchlineRequest.cs
@@ -1,0 +1,13 @@
+﻿public struct PlayerPunchlineRequest
+{
+  public PlayerPunchlineRequest(string setup, string punchlineTemplate, string jokeId)
+  {
+    Setup = setup;
+    PunchlineTemplate = punchlineTemplate;
+    JokeId = jokeId;
+  }
+  public string Setup;
+  public string PunchlineTemplate;
+  public string JokeId;
+}
+
