+++ SharedPackages/Assets/Scripts/PlayerPunchlineResponse.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5ca2b6e18ce92b242958e947be4a0898
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
