+++ SharedPackages/Assets/Scripts/PlayerSetupResponse.cs
@@ -1,0 +1,13 @@
+﻿
+// Round Messages
+public struct PlayerSetupResponse
+{
+  public PlayerSetupResponse(string setup, string jokeId)
+  {
+    Setup = setup;
+    JokeId = jokeId;
+  }
+  public string Setup;
+  public string JokeId;
+}
+
